requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.0
fsspec==2025.3.2
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pillow_heif==0.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
roboflow==1.1.63
ruff==0.11.9
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
